chore: resolution error when trying module fallback with ts source imports
