+++ packages/uikit-react-native-foundation/src/components/Image/FastImageInternal.tsx
@@ -1,0 +1,9 @@
+import type { FastImageProps } from 'react-native-fast-image';
+
+let FastImageInternal: (props: FastImageProps) => JSX.Element | null = () => null;
+
+try {
+  FastImageInternal = require('react-native-fast-image') as (props: FastImageProps) => JSX.Element;
+} catch {}
+
+export default FastImageInternal;
